pyproject.toml
[tool.poetry]
name = "lex_upload"
version = "0.1.0"
description = ""
authors = ["trasba <trasba>"]
packages = [
    { include = "lex_upload" },
    { include = "lex_upload/lib/*.py" },
]

[tool.poetry.dependencies]
python = "^3.7"
"pdfminer.six" = "^20220319"
requests = "^2.27.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "piwheels"
url = "https://www.piwheels.org/simple"
secondary = true